Move QueryParameters to utils dir, as this is not filter specific
